add `sha256` package with simplified names, use it in test
